game + physics + shipyard + terrain gen: vessel sim (fuel, throttle, attitude autopilot, Δv), climate/biome stages
physics: AttitudeState, ShipParameters, ControlInput on Simulation;
ship mass tracked through burns and clamped at dry mass; Tsiolkovsky-
aware burn_duration capped by propellant. KeplerianPropagator's RK4
burn now carries attitude; trajectory prediction stays attitude-free.

game: fuel.rs gates throttle (KSP keys) on per-tank reactant drain
and reconciles tank pools to the simulation's ground-truth burn each
frame. engine.rs publishes EngineThrust per part with a temporary
tint visual. navigation.rs adds an autopilot for Stability /
Prograde / Retrograde / Normal / Radial / Target / ManeuverNode
modes. Apoapsis/periapsis markers in flight_plan_view. Dual cameras
(map vs ship) with Free/Orbital ShipCameraMode and CoM-pivoted
CameraTargetOffset. Opt-in mesh picking so body meshes stop
absorbing maneuver-arrow rays.

shipyard: stats.rs derives ShipStats (dry mass, thrust, Δv,
per-resource totals) from a blueprint. Δv is Tsiolkovsky on the
mass-bearing reactant pool that runs out first; electricity is a
continuous draw, not a reactant.

terrain gen: Climate stage (zonal precipitation + upwind
continentality + orographic gain / rain shadow), Biomes stage
(Whittaker first-match partition with per-biome iron_visibility),
expanded surface_materials, paint_biomes overhaul.
solar_system.ron switches Thalos to the new pipeline.

Plus: planet_impostor shader work and CLAUDE.md refresh for the new
physics architecture.

+++ crates/terrain_gen/src/stages/megabasin.rs
@@ -161,23 +161,23 @@ impl Stage for Megabasin {
                             // rim position is not a smooth circle but a
                             // fractally perturbed contour.
                             let warp_lo = fbm3(
-                                dir.x as f64 * 2.2,
-                                dir.y as f64 * 2.2,
-                                dir.z as f64 * 2.2,
-                                b.seed,
+                                dir.x * 2.2,
+                                dir.y * 2.2,
+                                dir.z * 2.2,
+                                b.seed as u32,
                                 4,
                                 0.55,
                                 2.0,
-                            ) as f32;
+                            );
                             let warp_hi = fbm3(
-                                dir.x as f64 * 7.5,
-                                dir.y as f64 * 7.5,
-                                dir.z as f64 * 7.5,
-                                b.seed ^ 0x27F1_AC63_5B2E_94D8,
+                                dir.x * 7.5,
+                                dir.y * 7.5,
+                                dir.z * 7.5,
+                                (b.seed ^ 0x27F1_AC63_5B2E_94D8) as u32,
                                 3,
                                 0.5,
                                 2.2,
-                            ) as f32;
+                            );
                             let warp = warp_lo * 0.25 + warp_hi * 0.10;
                             let t = t_raw / (1.0 + warp);
 
@@ -221,23 +221,23 @@ impl Stage for Megabasin {
                             let azimuth = ty.atan2(tx);
 
                             let warp_lo = fbm3(
-                                dir.x as f64 * 2.0,
-                                dir.y as f64 * 2.0,
-                                dir.z as f64 * 2.0,
-                                b.seed ^ 0xC3A9_17B5_FD21_48C7,
+                                dir.x * 2.0,
+                                dir.y * 2.0,
+                                dir.z * 2.0,
+                                (b.seed ^ 0xC3A9_17B5_FD21_48C7) as u32,
                                 4,
                                 0.55,
                                 2.1,
-                            ) as f32;
+                            );
                             let warp_hi = fbm3(
-                                dir.x as f64 * 6.5,
-                                dir.y as f64 * 6.5,
-                                dir.z as f64 * 6.5,
-                                b.seed ^ 0x5E71_B2F9_3D88_9156,
+                                dir.x * 6.5,
+                                dir.y * 6.5,
+                                dir.z * 6.5,
+                                (b.seed ^ 0x5E71_B2F9_3D88_9156) as u32,
                                 3,
                                 0.5,
                                 2.3,
-                            ) as f32;
+                            );
                             let warp = warp_lo * 0.35 + warp_hi * 0.12;
                             let t = t_raw / (1.0 + warp);
